requirements.txt
torch
torchvision
pillow
numpy
huggingface_hub
transformers>=4.39.1